Use LCG to randomize pixel channel choices
The LCG is a pseudorandom number generator that can create a perfect
full cycle through the given range. This means that each pixel channel
will be chosen exactly once in a worst-case scenario where 100% of the
RGB pixel channels are required to store the hidden message, without
needing to keep track of the pixels already chosen.