Fix to directory structure
Importing package.json to get the version number inside the CLI program
caused the dist tree to shift from dist/index.js to dist/src/index.js.
This corrects the package.json "main" and "bin" entries to reflect that.

Yes it's a bit strange to have dist/package.json and dist/src in the
published bundle, but this keeps things clean from a maintenance
perspective, and people rarely inspect the bundle contents closely, esp
when things are working properly.

This also takes advantage of pnpm run --recursive to avoid needing an
all.sh to run scripts for each package.

Also, with regard to version numbers, let's use the highest version
number available in any sub-package for the git tag and the outer
package version number.