shell-ast: Experimenting with a much simpler AST
This configuration enables tuple typing for arguments, rather than union
types. It also requires all the properties set but provides some helper
functions to complete them as needed.

This is the ultimate goal of the project. The idea is to provide a
generic AST description of a CLI command and then adapt it to various
libraries such as commander.js and omelette.

I would just switch to yargs but holy crap that interface is sucky. It
definitely doesn't follow the principle of least surprise.